feat(@unional/devpkg-node): accept other test matches
Before this change,
jest will pick up `spec` and `test` using the presets of `ts-jest`.

After the change,
it is expanded to `spec`, `test`, `unit`, `integrate`, `system`, `accept`.

Ideally, use `spec` as acceptance tests.
But since it is so wide spread to use `spec` for everything,
you can also use `accept` for that purpose.
With these test categories,
you can gradually remove the usage of generic `spec` and `test`.